Moved shared view model methods to a module.
