+++ app/view_models/volca_share/shared.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module VolcaShare
+  module Shared
+    def checked?(field)
+      return { checked: true } if model.send(field)
+      {}
+    end
+
+    def lit?(field)
+      return 'lit' if model.send(field)
+      'unlit'
+    end
+  end
+end
